+++ modules/tinymce/src/core/main/ts/caret/ClosestCaretCandidate.ts
@@ -51,22 +51,42 @@ const clientInfo = (rect: NodeClientRect, clientX: number): FakeCaretInfo => {
   };
 };
 
-const horizontalDistance: DistanceFn = (rect, x, _y) => Math.min(Math.abs(rect.left - x), Math.abs(rect.right - x));
+const horizontalDistance: DistanceFn = (rect, x, _y) => x > rect.left && x < rect.right ? 0 : Math.min(Math.abs(rect.left - x), Math.abs(rect.right - x));
 
 const closestChildCaretCandidateNodeRect = (children: ChildNode[], clientX: number, clientY: number): Optional<NodeClientRect> => {
-  const findClosestCaretCandidateNodeRect = (rects: NodeClientRect[], distance: DistanceFn): Optional<NodeClientRect> => {
-    return Arr.findMap(
-      Arr.sort(rects, (r1, r2) => distance(r1, clientX, clientY) - distance(r2, clientX, clientY)),
-      (rect) => {
-        if (CaretCandidate.isCaretCandidate(rect.node)) {
-          return Optional.some(rect);
-        } else if (NodeType.isElement(rect.node)) {
-          return closestChildCaretCandidateNodeRect(Arr.from(rect.node.childNodes), clientX, clientY);
-        } else {
-          return Optional.none();
+  const caretCandidateRect = (rect: NodeClientRect) => {
+    if (CaretCandidate.isCaretCandidate(rect.node)) {
+      return Optional.some(rect);
+    } else if (NodeType.isElement(rect.node)) {
+      return closestChildCaretCandidateNodeRect(Arr.from(rect.node.childNodes), clientX, clientY);
+    } else {
+      return Optional.none();
+    }
+  };
+
+  // If an element and a text node has nearly equal distance then favor the text node over the element to make it easier to select text
+  // since setting the selection range will cancel any text select operation.
+  const getClosestTextNode = (rects: NodeClientRect[], distance: DistanceFn) => {
+    if (rects.length >= 2) {
+      const r1 = caretCandidateRect(rects[0]).getOr(rects[0]);
+      const r2 = caretCandidateRect(rects[1]).getOr(rects[1]);
+      const deltaDistance = Math.abs(distance(r1, clientX, clientY) - distance(r2, clientX, clientY));
+
+      if (deltaDistance < 2) {
+        if (NodeType.isText(r1.node)) {
+          return Optional.some(r1);
+        } else if (NodeType.isText(r2.node)) {
+          return Optional.some(r2);
         }
       }
-    );
+    }
+
+    return Optional.none();
+  };
+
+  const findClosestCaretCandidateNodeRect = (rects: NodeClientRect[], distance: DistanceFn): Optional<NodeClientRect> => {
+    const sortedRects = Arr.sort(rects, (r1, r2) => distance(r1, clientX, clientY) - distance(r2, clientX, clientY));
+    return getClosestTextNode(sortedRects, distance).orThunk(() => Arr.findMap(sortedRects, caretCandidateRect));
   };
 
   const [ horizontalRects, verticalRects ] = splitRectsPerAxis(getClientRects(children), clientY);
@@ -87,9 +107,7 @@ const traverseUp = (rootElm: SugarElement<Node>, scope: SugarElement<Element>, c
       }
     ).orThunk(() => {
       const parent = Compare.eq(scope, rootElm) ? Optional.none() : Traverse.parentElement(scope);
-      return parent
-        .filter((newScope) => !Compare.eq(newScope, rootElm))
-        .bind((newScope) => helper(newScope, Optional.some(scope)));
+      return parent.bind((newScope) => helper(newScope, Optional.some(scope)));
     });
   };